Refactor the code completely
