+++ src/commonComponents/PrivateRoute.jsx
@@ -4,7 +4,7 @@ import { Navigate, useLocation } from 'react-router-dom';
 import CircularProgress from '@mui/material/CircularProgress';
 import Box from '@mui/material/Box';
 
-import { useAuth } from '../../contexts/AuthContext';
+import { useAuth } from '../contexts/AuthContext';
 
 export default function PrivateRoute({ children }) {
   const { isAuthenticated, loading } = useAuth();
