+++ src/hooks/useDebounce.js
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from 'react';
 
-export function useDebounce(value, delay = 300) {
+export default function useDebounce(value, delay = 300) {
   const [debouncedValue, setDebouncedValue] = useState(value);
 
   useEffect(() => {
