+++ src/services/feedbackService.js
@@ -1,13 +1,12 @@
 // src/services/feedbackService.js
 
-import { supabase } from './supabaseClient';
+import supabase from './supabaseClient';
 
-export const fetchUserEmail = async () => {
-  const { data } = await supabase.auth.getUser();
-  return data?.user?.email || null;
-};
+export const insertFeedback = async ({ message, user }) => {
+  if (!userData?.user?.id) return false;
 
-export const insertFeedback = async ({ message, email }) => {
+  const email = userData.user.email;
+
   const { error } = await supabase.from('feedback').insert([{ message, email }]);
   if (error) {
     console.error('Feedback insert error:', error);
