+++ src/services/transferService.js
@@ -1,0 +1,8 @@
+// src/services/transferService.js
+
+export const getTransfers = async () => {
+  // let's work on this later
+
+  const data = [];
+  return data || [];
+};
